improve `utils.sh` with more helper functions

+++ utils.sh
@@ -19,53 +19,107 @@ fi
 # TODO add check that UTILS_ROOT is absolute
 
 
-# creates a project from a template
-create-project(){
-  [[ "$#" == 3 ]] || {
-    echo; echo "usage: create-project TEMPLATE PROJ_DIR PROJ_NAME"
-    echo "available TEMPLATEs are: zynq7000, ultrascale, zedboard, coraz7"; echo
-    return 0
-  }
+#
+### Helper functions ###
+#
 
-  local TEMPLATE="$1"
-  local PROJ_DIR="$2"
-  local PROJ_NAME="$3"
-  local SCRIPT="$UTILS_ROOT/templates/$TEMPLATE.tcl"
+# Args:
+# $1 level
+# $2 message
+# $3 return code, if termination is desired
+log(){
+  local RED='\e[31m'
+  local GREEN='\e[32m'
+  local BOLD='\e[1m'
+  local RESET='\e[0m'
+  case "$1" in
+    info)
+      echo -e "${BOLD}[${GREEN}INFO${RESET}${BOLD}]${RESET}  ${2//\\n/\n        }"
+    ;;
+    error)
+      echo -e "${BOLD}[${RED}ERROR${RESET}${BOLD}]${RESET} ${2//\\n/\n        }"
+    ;;
+    usage)
+      echo -e "usage: $2\n"
+      for line in "${@:3}"
+      do
+        echo -e "$line"
+      done
+      exit 0
+    ;;
+    *)
+      log error "unknown loglevel $1" 1
+    ;;
+  esac
 
-  [[ -f "$SCRIPT" ]] || {
-    echo; echo "The template script '$SCRIPT' could not be found."; echo
-    return 1
-  }
-
-  [[ -d "$PROJ_DIR" ]] || {
-    echo; echo "The project directory '$PROJ_DIR' does not exist; creating it"; echo
-    mkdir --parent -- "$PROJ_DIR"
-  }
-
-  if [[ -d "$PROJ_DIR/$PROJ_NAME" ]]
+  if [[ "$#" -ge 3 ]]
   then
-    echo "Delete existing project..."; echo
-    # rm --recursive --force -- "$PROJ_DIR/$PROJ_NAME/"*
+    exit "$3"
+  fi
+}
+
+# ensures that $1 is a file.
+#
+# arguments: file|dir <variable name> <path>
+#
+# behavior codes
+# exit 1: terminates with return code 1 if $1 is not a valid variable name
+# exit 127: terminates with return code 127 if $2 is not a file/dir
+# exit 0: stores the path to $2 in a variable with the name of $1
+_ensure-exists(){
+  local TYPE="$1"
+  local VARIABLE_NAME="$2"
+  local PATH="$3"
+
+  # check that variable name is a valid variable name
+  [[ "$VARIABLE_NAME" =~ ^[_[:alpha:]][_[:alpha:][:digit:]]*$ ]] || {
+    log error "the variable name '$1' is not valid" 1
+  }
+
+  if [[ "$TYPE" == file ]] && [[ ! -f "$PATH" ]]
+  then
+    log error "the $VARIABLE_NAME $TYPE '$PATH' could not be found" 127
   fi
 
-  mkdir --parent -- "$PROJ_DIR/$PROJ_NAME/"{workspace,constr,hdl,scripts}
+  if [[ "$TYPE" == dir ]] && [[ ! -d "$PATH" ]]
+  then
+    log error "the $VARIABLE_NAME $TYPE '$PATH' could not be found" 127
+  fi
 
-  cp "$UTILS_ROOT/templates/.gitignore" "$PROJ_DIR/$PROJ_NAME"
-  echo -e "# README for the '$PROJ_NAME' FPGA project\n" > "$PROJ_DIR/$PROJ_NAME/README.md"
+  eval "$VARIABLE_NAME='$PATH'"
+}
 
-  update-wrapper "$PROJ_DIR" "$PROJ_NAME"
+# checks that $1 does not exist, terminating with return code 1 after printing $2 on violation
+_ensure-exists-not(){
+  if [[ -e "$1" ]]
+  then
+    log error "'$1' already exists\n$2" 1
+  fi
+}
 
-  vivado -nolog -nojournal -mode batch -source "$SCRIPT" -tclargs "$PROJ_DIR" "$PROJ_NAME"
-  echo; echo "Vivado project created."
+# for each argument, ensure it  exists and is a directory, creating it if necessary
+# fail if any argument already exists but is not a dir
+_mkdir-if-missing(){
+  for dir in "$@"
+  do
+    # it is a dir, everything fine
+    [[ -d "$dir" ]] && continue
+
+    # it does not exist, create it
+    [[ -e "$dir" ]] || {
+      log info "the dir '$dir' does not exist; creating it"
+      mkdir --parent -- "$dir"
+      continue
+    }
+
+    # it does exist and is not a dir
+    log error "'$dir' exists but is not a directory, terminating" 1
+  done
 }
 
 # updates wrapper file of Vivado project
-update-wrapper()
-{
-  [[ "$#" == 2 ]] || {
-    echo; echo "usage: update-wrapper PROJ_DIR PROJ_NAME"; echo
-    return 0
-  }
+_update-wrapper(){
+  [[ "$#" == 2 ]] || log usage "update-wrapper PROJ_DIR PROJ_NAME"
 
   local PROJ_DIR="$1"
   local PROJ_NAME="$2"
@@ -73,64 +127,59 @@ update-wrapper()
   python3 "$UTILS_ROOT/templates/set_update_wrapper.py" "$PROJ_DIR/$PROJ_NAME" "$PROJ_NAME"
 }
 
-# stores a Vivado project
-store(){
-  [[ "$#" == 1 ]] || {
-    echo; echo "usage: store PROJ_DIR"; echo
-    return 0
-  }
-
-  local PROJ_DIR="$1"
-  local WORKSPACE_DIR="$PROJ_DIR/workspace"
-  local SCRIPT="$UTILS_ROOT/tcl-lib/store.tcl"
-
-  [[ -d "$PROJ_DIR" ]] || {
-    echo; echo "The project directory '$PROJ_DIR' does not exist."; echo
-    return 127
-  }
-
-  [[ -d "$WORKSPACE_DIR" ]] || {
-    echo; echo "The workspace directory '$WORKSPACE_DIR' does not exist."
-    echo "Please try to restore the workspace before storing it."; echo
-    return 127
-  }
-
-  vivado -nolog -nojournal -mode batch -source "$SCRIPT" "$WORKSPACE_DIR/"*.xpr
+# cleans up build artifacts from viviad & vitis
+_clean_up(){
   [[ -d ".Xil/" ]] &&  rm --recursive .Xil/
 }
 
+
+#
+### Functions made to be used externally ###
+#
+
+# creates a project from a template
+create-project(){
+  [[ "$#" == 3 ]] || log usage "create-project TEMPLATE PROJ_DIR PROJ_NAME" \
+    "available TEMPLATEs are: zynq7000, ultrascale, zedboard, coraz7"
+
+  _ensure-exists file SCRIPT "$UTILS_ROOT/templates/$1.tcl"
+  local PROJ_DIR="$2"
+  local PROJ_NAME="$3"
+  _ensure-exists-not "$PROJ_DIR/$PROJ_NAME" "Please delete existing project..."
+  _mkdir-if-missing "$PROJ_DIR/$PROJ_NAME/"{workspace,constr,hdl,scripts}
+
+  cp "$UTILS_ROOT/templates/.gitignore" "$PROJ_DIR/$PROJ_NAME"
+  echo -e "# README for the '$PROJ_NAME' FPGA project\n" > "$PROJ_DIR/$PROJ_NAME/README.md"
+
+  _update-wrapper "$PROJ_DIR" "$PROJ_NAME"
+
+  vivado -nolog -nojournal -mode batch -source "$SCRIPT" -tclargs "$PROJ_DIR" "$PROJ_NAME"
+}
+
+
+# stores a Vivado project
+store(){
+  [[ "$#" == 1 ]] || log usage "usage: store PROJ_DIR"
+
+  _ensure-exists file PROJ_FILE "$1/workspace"/*.xpr
+  _ensure-exists file SCRIPT "$UTILS_ROOT/tcl-lib/store.tcl"
+
+  vivado -nolog -nojournal -mode batch -source "$SCRIPT" "$PROJ_FILE"
+}
+
+
 # restores a project from a template script
 restore(){
-  [[ "$#" == 1 ]] || {
-    echo; echo "usage: restore PROJ_DIR"; echo
-    return 0
-  }
+  [[ "$#" == 1 ]] || log usage "restore PROJ_DIR"
 
-  local PROJ_DIR="$1"
-  local SCRIPT="$PROJ_DIR/scripts/restore.tcl"
-
-  [[ -d "$PROJ_DIR" ]] || {
-    echo; echo "The project directory '$PROJ_DIR' does not exist."; echo
-    return 127
-  }
-
-  [[ -f "$SCRIPT" ]] || {
-    echo; echo "The restore script '$SCRIPT' could not be found."
-    echo "Please try store the project before restoring it."; echo
-    return 1
-  }
-
-   if [ -d "$PROJ_DIR/workspace" ]
-   then
-    echo; echo "The workspace '$PROJ_DIR/workspace/' already exists."
-    echo "Please delete it manually before restoring the project."; echo
-    return 1
-  fi
+  _ensure-exists dir PROJ_DIR "$1"
+  _ensure-exists "$PROJ_DIR/scripts/restore.tcl"
+  _ensure-exists-not "$PROJ_DIR/workspace" "The workspace '$PROJ_DIR/workspace/' already exists" \
+    "Please delete it manually before restoring the project."
 
   pushd "$PROJ_DIR" || return 1
   vivado -nolog -nojournal -mode batch -source "$SCRIPT"
-  [[ -d ".Xil/" ]]
-    rm -r .Xil/
+  [[ -d ".Xil/" ]] && rm -r .Xil/
 
   # shellcheck disable=SC2164
   popd
@@ -139,60 +188,35 @@ restore(){
 # generates the harwdare configuration for a fpga project
 # Note: WORKSPACE_DIR must contain a subdir `workspace`
 generate-hw-config(){
-  [[ "$#" == 1 ]] || {
-    echo; echo "usage: generate-hw-config PROJ_DIR"; echo
-    return 0
-  }
+  [[ "$#" == 1 ]] || log usage "generate-hw-config PROJ_DIR"
 
-  local PROJ_DIR="$1"
-
-  [[ -d "$PROJ_DIR/workspace" ]] || {
-    echo; echo "The workspace directory '$PROJ_DIR/workspace' does not exist."
-    echo "Please try to restore the project before generating the hardware configuration."; echo
-    return 127
-  }
-
-  mkdir --parent -- "$PROJ_DIR/hardware_configuration"
+  _ensure-exists dir  PROJ_DIR "$1"
+  _ensure-exists file PROJ_FILE "$PROJ_DIR/workspace/"*.xpr
+  _mkdir-if-missing "$PROJ_DIR/hardware_configuration"
 
   vivado -nolog -nojournal -mode batch \
-    -source "$UTILS_ROOT/tcl-lib/vivado_all.tcl" \
-    "$PROJ_DIR/workspace/"*.xpr -tclargs "$PROJ_DIR/hardware_configuration"
+    -source "$UTILS_ROOT/tcl-lib/vivado_all.tcl" "$PROJ_FILE" \
+    -tclargs "$PROJ_DIR/hardware_configuration"
 }
 
 # generates the bootloader for a fpga project
 build-bootloader(){
-  [[ "$#" == 2 ]] || {
-    echo "usage: build-bootloader PLATFORM PROJ_DIR"
-    echo "available platforms are: zynq7000, zedboard, ultrascale, coraz7"; echo
-    return 0
-  }
+  [[ "$#" == 2 ]] || log usage "build-bootloader PLATFORM PROJ_DIR" \
+    "available platforms are: zynq7000, zedboard, ultrascale, coraz7"
 
   local PLATFORM="$1"
-  local PROJ_DIR="$2"
-
-  [[ -f "$UTILS_ROOT/tcl-lib/vitis_platform_$PLATFORM.tcl" ]] || {
-    echo; echo "Unknown platform $PLATFORM."; echo
-    return 1
-  }
-
-  [[ -d "$PROJ_DIR" ]] || {
-    echo; echo "PROJ_DIR '$OUT_DIR' does not exist."; echo
-    return 127
-  }
-
-  [[ -f "$PROJ_DIR/hardware_configuration/hw_export.xsa" ]] || {
-    echo; echo "Hardware configuration does not exist."; echo
-    return 127
-  }
+  _ensure-exists dir PROJ_DIR "$2"
+  _ensure-exists file SCRIPT "$UTILS_ROOT/tcl-lib/vitis_platform_$PLATFORM.tcl"
+  _ensure-exists file XSA "$PROJ_DIR/hardware_configuration/hw_export.xsa"
 
   if [[ -d "$PROJ_DIR/bootloader_configuration" ]] || [[ -d "$PROJ_DIR/bootloader_export" ]]
   then
     rm --recursive --force -- "$PROJ_DIR/"{bootloader_configuration,bootloader_export}
   fi
 
-  mkdir --parent -- "$PROJ_DIR/"{bootloader_configuration,bootloader_export}
+  _mkdir-if-missing "$PROJ_DIR/"{bootloader_configuration,bootloader_export}
 
-  xsct "$UTILS_ROOT/tcl-lib/vitis_platform_$PLATFORM.tcl" "$PROJ_DIR/bootloader_configuration" "$PROJ_DIR/hardware_configuration/"*.xsa
+  xsct "$SCRIPT" "$PROJ_DIR/bootloader_configuration" "$XSA"
 
   if [[ "$PLATFORM" == "ultrascale" ]]
   then
@@ -224,134 +248,42 @@ build-bootloader(){
 
 # boot an image via jtag
 jtag-boot(){
-  [[ "$#" == 3 ]] || {
-    echo "usage: jtag-boot HW_PLATFORM PROJECT_DIR ELF"
-    echo "available hw platforms are: zynq7000, coraz7, zedboard, ultrascale"; echo
-    return 0
-  }
+  [[ "$#" == 3 ]] || log usage "jtag-boot HW_PLATFORM PROJ_DIR ELF" \
+    "available hw platforms are: zynq7000, coraz7, zedboard, ultrascale"
 
   local HW_PLATFORM="$1"
-  local PROJECT_DIR="$2"
-  local ELF="$3"
+  _ensure-exists file SCRIPT "$UTILS_ROOT/tcl-lib/jtag_boot/${HW_PLATFORM}_init.tcl"
+  _ensure-exists dir PROJ_DIR "$2"
+  _ensure-exists file XSA "$(find "$PROJ_DIR" -name hw_export.xsa | head -n 1)"
+  _ensure-exists file ELF "$3"
 
-  [[ -f "$UTILS_ROOT/tcl-lib/jtag_boot/${HW_PLATFORM}_init.tcl" ]] || {
-    echo "unknown HW_PLATFORM $HW_PLATFORM"
-    return 1
-  }
+  local XSA_TMP_DIR
+  XSA_TMP_DIR=$(mktemp --directory -- "/tmp/$HW_PLATFORM-xsa-XXXXXXXXXX")
+  unzip -d "$XSA_TMP_DIR" -qq -- "$XSA"
+  log info "extracted XSA to $XSA_TMP_DIR"
 
-  local BITSTREAM
-  BITSTREAM=$(find "$PROJECT_DIR" -name bitstream_export.bit | head -n 1)
-  if [[ -f "$BITSTREAM" ]]
-  then
-    echo "found bitstream file '$BITSTREAM'"
-  else
-    echo "bitstream file '$BITSTREAM' not found"
-    return 127
-  fi
+  _ensure-exists file BITSTREAM "$XSA_TMP_DIR/hw_export.bit"
 
-  local XSA
-  XSA=$(find "$PROJECT_DIR" -name hw_export.xsa | head -n 1)
-  if [[ -f "$XSA" ]]
-  then
-    echo "found xsa file '$XSA'"
-  else
-    echo "xsa file '$XSA' not found"
-    return 127
-  fi
+  # the different HW platforms differ in their initialization, thus a switch case
+  case "$HW_PLATFORM" in
+    "zynq7000"|"zedboard"|"coraz7")
+      _ensure-exists file PS7 "$XSA_TMP_DIR/ps7_init.tcl"
 
-  [[ -f "$ELF" ]] || {
-    echo "ELF file '$ELF' not found"
-    return 127
-  }
+      xsct "$SCRIPT" "$PS7" "$BITSTREAM" "$XSA" "$ELF"
+    ;;
 
-  if [[ "$HW_PLATFORM" == "zynq7000" ]]
-  then
-    # needed: ps7_init.tcl
-    # bitstream
-    # hw_export.xsa
-    # elf
-    local PS7_FILE
-    PS7_FILE=$(find "$PROJECT_DIR" -name ps7_init.tcl | head -n 1)
-    if [[ -f "$PS7_FILE" ]]
-    then
-      echo "found ps7_init file '$PS7_FILE'"
-    else
-      echo "PS7 file '$PS7_FILE' not found"
-      return 127
-    fi
+    "ultrascale")
+      _ensure-exists file PMUFW "$(find "$PROJ_DIR" -name pmufw_ultrascale.elf | head -n 1)"
+      _ensure-exists file FSBL "$(find "$PROJ_DIR" -name fsbl_ultrascale.elf | head -n 1)"
+      _ensure-exists file ZYNQMP_UTILS "$(find "$UTILS_ROOT/tcl-lib" -name zynqmp_utils.tcl | head -n 1)"
 
-    xsct "$UTILS_ROOT/tcl-lib/jtag_boot/${HW_PLATFORM}_init.tcl" "$PS7_FILE" "$BITSTREAM" "$XSA" "$ELF"
+      xsct "$SCRIPT" "$BITSTREAM" "$XSA" "$PMUFW" "$FSBL" "$ZYNQMP_UTILS" "$ELF"
+    ;;
 
-  elif [[ "$HW_PLATFORM" == "ultrascale" ]]
-  then
-    # needed: zynqmp_utils.tcl
-    # bitstream
-    # hw_export.xsa
-    # pmufw_ultrascale.elf
-    # fsbl_ultrascale.elf
-    # zynqmp_utils.tcl
-    # elf
-    local PMUFW_FILE
-    PMUFW_FILE=$(find "$PROJECT_DIR" -name pmufw_ultrascale.elf | head -n 1)
-    if [[ -f "$PMUFW_FILE" ]]
-    then
-      echo "found pmufw_ultrascale file '$PMUFW_FILE'"
-    else
-      echo "pmufw_ultrascale file '$PMUFW_FILE' not found"
-      return 127
-    fi
+    *)
+      log error "unsupported platform $HW_PLATFORM" 1
+    ;;
+  esac
 
-    local FSBL_FILE
-    FSBL_FILE=$(find "$PROJECT_DIR" -name fsbl_ultrascale.elf | head -n 1)
-    if [[ -f "$FSBL_FILE" ]]
-    then
-      echo "found fsbl_ultrascale file '$FSBL_FILE'"
-    else
-      echo "fsbl_ultrascale file '$FSBL_FILE' not found"
-      return 127
-    fi
-
-    local ZYNQMP_UTILS
-    ZYNQMP_UTILS=$(find "$UTILS_ROOT/tcl-lib" -name zynqmp_utils.tcl | head -n 1)
-    if [[ -f "$ZYNQMP_UTILS" ]]
-    then
-      echo "found zynqmp_utils file '$ZYNQMP_UTILS'"
-    else
-      echo "zynqmp_utils file '$ZYNQMP_UTILS' not found"
-      return 127
-    fi
-
-    xsct "$UTILS_ROOT/tcl-lib/jtag_boot/${HW_PLATFORM}_init.tcl" "$BITSTREAM" "$XSA" "$PMUFW_FILE" "$FSBL_FILE" "$ZYNQMP_UTILS" "$ELF"
-
-  elif [[ "$HW_PLATFORM" == "zedboard" ]]
-  then
-    # needed: ps7_init.tcl
-    # bitstream
-    # hw_export.xsa
-    # elf
-    local PS7_FILE
-    PS7_FILE=$(find "$PROJECT_DIR" -name ps7_init.tcl | head -n 1)
-    if [[ -f "$PS7_FILE" ]]
-    then
-      echo "found ps7_init file '$PS7_FILE'"
-    else
-      echo "PS7 file '$PS7_FILE' not found"
-      return 127
-    fi
-
-    xsct "$UTILS_ROOT/tcl-lib/jtag_boot/${HW_PLATFORM}_init.tcl" "$PS7_FILE" "$BITSTREAM" "$XSA" "$ELF"
-
-  elif [[ "$HW_PLATFORM" == "coraz7" ]]
-  then
-    # needed: ps7_init.tcl
-    # bitstream
-    # hw_export.xsa
-    # elf
-    # TODO implement the rest
-    echo "currently unsupported"
-    return 1
-  else
-    echo "unsupported platform $HW_PLATFORM"
-    return 1
-  fi
+  rm --recursive -- "$XSA_TMP_DIR"
 }
